% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleChangeOfVariables-package.r
\docType{package}
\name{nimbleChangeOfVariables}
\alias{nimbleChangeOfVariables}
\title{nimbleChangeOfVariables}
\description{
A collection of nimble functions for sampling common probability distributions on unbounded scales.
}
\section{Transformed probability distributions for unbounded sampling in nimble.}{
NA
}

\seealso{
\url{https://r-nimble.org/}
}
